Create guest_net_timeout variable.
When using this module to deploy within our infrastructure we
encountered time out errors while waiting on the deployment of the AVI
appliances.

Increasing the value of the wait_for_guest_net_timeout attribute of the
vsphere_virtual_machine.avi_controller resource from the default value
of 1 to 2 solved this problem for us.

This change introduces the guest_net_timeout variable to the module with
to allow this attribute to be configured on a case-by-case basis.  The
default value is still 1.