Make se_mgmt_portgroup optional.
The ansible playbook generated by the template
files/avi-vsphere-all-in-one-play.yml.tpl expects the value of
se_mgmt_portgroup to not be null even when we do not intend to configure
the se_mgmt_network.  This change removes that variable from the
generated playbook when configure_se_mgmt_network is false.